Created About page.
Jekyll doesn’t handle pages, so as I work around I created a page
layout so we don’t have to worry about updating the header/nav/footer
on all pages in the future. The only downside is that the page shows up
in the posts feed on the homepage.